// frontend/src/proto/random_pb.d.ts
  export type AsObject = {
    name: string,
    avatarUrl: string,
  }
// ...
export class User extends jspb.Message {
  getId(): number;
  setId(value: number): User;

  getName(): string;
  setName(value: string): User;

  getStatus(): string;
  setStatus(value: string): User;

  getAvatarUrl(): string;
  setAvatarUrl(value: string): User;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): User.AsObject;
  static toObject(includeInstance: boolean, msg: User): User.AsObject;
  static serializeBinaryToWriter(message: User, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): User;
  static deserializeBinaryFromReader(message: User, reader: jspb.BinaryReader): User;
}

export namespace User {
  export type AsObject = {
    id: number,
    name: string,
    status: string,
    avatarUrl: string,
  }
}
// ...
export class StreamMessage extends jspb.Message {
  getUserId(): number;
  setUserId(value: number): StreamMessage;

  getUserAvatar(): string;
  setUserAvatar(value: string): StreamMessage;

  getMessage(): string;
  setMessage(value: string): StreamMessage;

  getUserName(): string;
  setUserName(value: string): StreamMessage;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): StreamMessage.AsObject;
  static toObject(includeInstance: boolean, msg: StreamMessage): StreamMessage.AsObject;
  static serializeBinaryToWriter(message: StreamMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): StreamMessage;
  static deserializeBinaryFromReader(message: StreamMessage, reader: jspb.BinaryReader): StreamMessage;
}

export namespace StreamMessage {
  export type AsObject = {
    userId: number,
    userAvatar: string,
    message: string,
    userName: string,
  }
}